Extract RelativePathSerializerMixin from MeshConfig
Move the paths_relative_to serialization logic into a reusable mixin
in celeri_util so it can be shared with Config in a follow-up.
